Refactor the ResourceGroup package to improve modularity and maintainability:
- Split schema related functionality into a new 'schema' subpackage
- Create a new 'crd' subpackage for CRD related operations
- Improve type inference and schema generation for CRDs.
- Enhance error handling and logging throughout the package.
- Update dependencies and remove unused functions
- Refactor utility functions for better reusability
