+++ internal/resourcegroup/resource_variables.go
@@ -1,0 +1,28 @@
+// Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"). You may
+// not use this file except in compliance with the License. A copy of the
+// License is located at
+//
+//	http://aws.amazon.com/apache2.0/
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package resourcegroup
+
+import "github.com/aws-controllers-k8s/symphony/internal/typesystem/parser"
+
+type ResourceVariableKind string
+
+const (
+	ResourceVariableKindStatic  ResourceVariableKind = "static"
+	ResourceVariableKindDynamic ResourceVariableKind = "dynamic"
+)
+
+type ResourceVariable struct {
+	parser.ExpressionField
+	Kind ResourceVariableKind
+}
