+++ internal/resourcegroup/schema/resolver.go
@@ -1,0 +1,50 @@
+// Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"). You may
+// not use this file except in compliance with the License. A copy of the
+// License is located at
+//
+//	http://aws.amazon.com/apache2.0/
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package schema
+
+import (
+	"k8s.io/apiextensions-apiserver/pkg/generated/openapi"
+	"k8s.io/apiserver/pkg/cel/openapi/resolver"
+	"k8s.io/client-go/discovery"
+	"k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/client-go/rest"
+)
+
+// NewCombinedResolver creates a new schema resolver that can resolve both core and client types.
+func NewCombinedResolver(clientConfig *rest.Config) (resolver.SchemaResolver, error) {
+	discoveryClient, err := discovery.NewDiscoveryClientForConfig(clientConfig)
+	if err != nil {
+		return nil, err
+	}
+
+	// ClientResolver is a resolver that uses the discovery client to resolve
+	// client types. It is used to resolve types that are not known at compile
+	// time a.k.a present in:
+	// https://github.com/kubernetes/apiextensions-apiserver/blob/master/pkg/generated/openapi/zz_generated.openapi.go
+	clientResolver := &resolver.ClientDiscoveryResolver{
+		Discovery: discoveryClient,
+	}
+
+	// CoreResolver is a resolver that uses the OpenAPI definitions to resolve
+	// core types. It is used to resolve types that are known at compile time.
+	coreResolver := resolver.NewDefinitionsSchemaResolver(
+		openapi.GetOpenAPIDefinitions,
+		scheme.Scheme,
+	)
+
+	// Combine the two resolvers to create a single resolver that can resolve
+	// both core and client types.
+	combinedResolver := coreResolver.Combine(clientResolver)
+	return combinedResolver, nil
+}
